Fixed user notice display, and refactored chat line views
