+++ public/src/js/chatview/ChatLine.jsx
@@ -3,14 +3,14 @@ import PropTypes from "prop-types";
 import { connect } from "react-redux";
 
 import ChannelLink from "../components/ChannelLink.jsx";
-import ChatBunchedEventsLine from "./ChatBunchedEventsLine.jsx";
-import ChatConnectionEventLine from "./ChatConnectionEventLine.jsx";
+import ChatBunchedEventsLine from "./chatline/ChatBunchedEventsLine.jsx";
+import ChatConnectionEventLine from "./chatline/ChatConnectionEventLine.jsx";
 import ChatContextView from "./ChatContextView.jsx";
-import ChatHighlightedLine from "./ChatHighlightedLine.jsx";
-import ChatMessageLine from "./ChatMessageLine.jsx";
-import ChatUserEventLine from "./ChatUserEventLine.jsx";
-import ChatUserNoticeLine from "./ChatUserNoticeLine.jsx";
-import LogLine from "./LogLine.jsx";
+import ChatHighlightedLine from "./chatline/ChatHighlightedLine.jsx";
+import ChatMessageLine from "./chatline/ChatMessageLine.jsx";
+import ChatUserEventLine from "./chatline/ChatUserEventLine.jsx";
+import ChatUserNoticeLinePrefix from "./chatline/ChatUserNoticeLinePrefix.jsx";
+import LogLine from "./chatline/LogLine.jsx";
 import { getChannelDisplayNameFromState } from "../lib/channelNames";
 import { prepareBunchedEvents } from "../lib/chatEvents";
 import { dateStamp, timeStamp } from "../lib/formatting";
@@ -30,6 +30,7 @@ class ChatLine extends PureComponent {
 			displayChannel,
 			highlight,
 			lineId,
+			message,
 			time,
 			type
 		} = this.props;
@@ -50,13 +51,19 @@ class ChatLine extends PureComponent {
 			(isNotice ? ` ${block}--notice` : "") +
 			(type === "connectionEvent" ? ` ${block}--connection` : "");
 
-		var content = null;
+		var content = null, prefix = null;
 
 		switch (type) {
 			case "msg":
 			case "action":
 			case "notice":
-				content = <ChatMessageLine {...this.props} key="content" />;
+			case "usernotice":
+				if (message) {
+					content = <ChatMessageLine {...this.props} key="content" />;
+				}
+				if (type === "usernotice") {
+					prefix = <ChatUserNoticeLinePrefix {...this.props} key="prefix" />;
+				}
 				break;
 			case "join":
 			case "part":
@@ -65,9 +72,6 @@ class ChatLine extends PureComponent {
 			case "kill":
 			case "mode":
 				content = <ChatUserEventLine {...this.props} key="content" />;
-				break;
-			case "usernotice":
-				content = <ChatUserNoticeLine {...this.props} key="content" />;
 				break;
 			case "events":
 				var { collapseJoinParts, ...bunchedProps } = this.props;
@@ -97,7 +101,7 @@ class ChatLine extends PureComponent {
 				break;
 		}
 
-		if (!content) {
+		if (!content && !prefix) {
 			return null;
 		}
 
@@ -124,12 +128,23 @@ class ChatLine extends PureComponent {
 			</time>
 		);
 
-		const outerContent = [
-			channelEl,
-			timeStampEl,
-			" ",
-			content
-		];
+		var outerContent;
+
+		if (content) {
+			outerContent = [
+				prefix,
+				channelEl,
+				timeStampEl,
+				" ",
+				content
+			];
+		}
+		else {
+			outerContent = [
+				prefix
+			];
+		}
+
 
 		const itemProps = {
 			className,
