+++ public/src/scss/chatview.scss
@@ -106,6 +106,11 @@
 		white-space: nowrap;
 	}
 
+	.line .prefix {
+		color: #999;
+		margin: 2px 0;
+	}
+
 	.msg__author a,
 	.userevent__target,
 	.userevent__target a,
@@ -248,19 +253,6 @@
 
 		&--highlight {
 			background-color: #fff0aa;
-		}
-	}
-
-	.usernotice {
-		> span,
-		> em {
-			display: block;
-			font-style: normal;
-		}
-
-		> em {
-			opacity: .75;
-			font-size: .75rem;
 		}
 	}
