+++ public/src/scss/darkmode.scss
@@ -43,6 +43,7 @@ body.darkmode {
 		.line__channel,
 		.line__channel a,
 		.line time,
+		.line .prefix,
 		.userevent,
 		.bunchedevents {
 			color: #777;
